Add README to describe the blogging process
We're all opinionated as to how we'd each like to blog. That's great,
but we need to adopt a process so that we're all on the same page as to
how to get content on the Gaslight blog. We'll document that here. As we
make changes, we need to keep this document up to date. It is the
canonical source for blogging at Gaslight.